+++ CandyBehavior.cs
@@ -21,6 +21,12 @@ namespace SCP956
 
         public bool pinataCandy = true;
 
+        public override void Start()
+        {
+            base.Start();
+            // TODO: Set node sub text here?
+        }
+
         public override void EquipItem()
         {
             base.EquipItem();
@@ -33,18 +39,25 @@ namespace SCP956
             if (configEnableCandyBag.Value) { playerHeldBy.equippedUsableItemQE = false; }
         }
 
-        public override void DiscardItem()
-        {
-            base.DiscardItem();
-            if (configEnableCandyBag.Value) { playerHeldBy.equippedUsableItemQE = false; }
-        }
-
         public override void ItemInteractLeftRight(bool right)
         {
             base.ItemInteractLeftRight(right);
+            if (right) { return; }
             if (configEnableCandyBag.Value)
             {
-                logger.LogDebug("ItemInteractLeftRight: " + right);
+                logger.LogDebug("Putting candy in bag");
+                playerHeldBy.DespawnHeldObject();
+
+                if (playerHeldBy.ItemSlots.Where(x => x != null && x.itemProperties.name == "BagOfCandyItem").FirstOrDefault() == null)
+                {
+                    NetworkHandler.Instance.SpawnItemServerRpc(playerHeldBy.actualClientId, "Candy Bag", 0, playerHeldBy.transform.position, Quaternion.identity, true);
+                }
+
+                GrabbableObject candyBagObj = playerHeldBy.ItemSlots.Where(x => x != null && x.itemProperties.name == "BagOfCandyItem").FirstOrDefault();
+                if (candyBagObj == null) { logger.LogError("Candy bag not found"); return; }
+
+                CandyBagBehavior candyBag = candyBagObj.gameObject.GetComponent<CandyBagBehavior>();
+                candyBag.CandyBag[itemProperties.itemName].Add(pinataCandy);
             }
         }
 
@@ -58,37 +71,6 @@ namespace SCP956
                     NetworkHandler.Instance.SpawnPinataNearbyServerRpc(StartOfRound.Instance.localPlayerController.transform.position);
                 }
             }
-
-            /*if (configEnableCandyBag.Value)
-            {
-                if (CandyBagBehavior.takingCandyOut) { CandyBagBehavior.takingCandyOut = false; return; }
-
-                CandyBagBehavior candyBag;
-                GrabbableObject candyBagObj = playerHeldBy.ItemSlots.Where(x => x != null && x.itemProperties.name == "BagOfCandyItem").FirstOrDefault();
-                GrabbableObject candy2Obj = playerHeldBy.ItemSlots.Where(x => x != null && x != this && CandyNames.Contains(x.itemProperties.itemName)).FirstOrDefault();
-
-                if (candyBagObj != null)
-                {
-                    candyBag = candyBagObj.gameObject.GetComponent<CandyBagBehavior>();
-                    candyBag.CandyBag[itemProperties.itemName].Add(pinataCandy);
-                    localPlayer.DespawnHeldObject();
-                }
-                else if (candy2Obj != null)
-                {
-                    string candy2Name = candy2Obj.itemProperties.itemName;
-                    bool candy2Pinata = candy2Obj.gameObject.GetComponent<CandyBehavior>().pinataCandy;
-                    DespawnItemInSlot(localPlayer.ItemSlots.IndexOf(candy2Obj));
-
-                    NetworkHandler.Instance.SpawnItemServerRpc(playerHeldBy.actualClientId, "Candy Bag", 0, playerHeldBy.transform.position, Quaternion.identity, true);
-                    candyBagObj = playerHeldBy.ItemSlots.Where(x => x != null && x.itemProperties.name == "BagOfCandyItem").FirstOrDefault();
-                    if (candyBagObj == null) { logger.LogError("Candy bag not found"); return; }
-
-                    candyBag = candyBagObj.gameObject.GetComponent<CandyBagBehavior>();
-                    candyBag.CandyBag[candy2Name].Add(candy2Pinata);
-                    candyBag.CandyBag[itemProperties.itemName].Add(pinataCandy);
-                    localPlayer.DespawnHeldObject();
-                }
-            }*/
         }
 
         public override void ItemActivate(bool used, bool buttonDown = true)
@@ -99,105 +81,104 @@ namespace SCP956
             {
                 logger.LogDebug("Eating candy");
                 playerHeldBy.DespawnHeldObject();
-                HUDManager.Instance.UIAudio.PlayOneShot(CandyCrunchsfx, 1f);
-                
-
-                if (!pinataCandy || PlayerAge >= 12)
-                {
-                    if (pinataCandy && (int)UnityEngine.Random.Range(0, 101) < configCandyDeathChance.Value)
-                    {
-                        playerHeldBy.KillPlayer(new Vector3(), true, CauseOfDeath.Unknown, 3);
-                        return;
-                    }
-
-                    if (!pinataCandy || configSecretLab.Value)
-                    {
-                        logger.LogDebug("Activating candy...");
-                        ActivateCandy(itemProperties.itemName);
-                    }
-                }
-                else if (configEnablePinata.Value)
-                {
-                    // TODO: Animation for player turning into SCP956 and bones crunching sound effects. Maybe spawn in as scavenger model and play animation to turn into scp956!
-                    playerHeldBy.KillPlayer(new Vector3(), false, CauseOfDeath.Unknown, 3);
-                    int index = RoundManager.Instance.currentLevel.Enemies.FindIndex(x => x.enemyType.enemyName == "SCP-956");
-                    RoundManager.Instance.SpawnEnemyOnServer(playerHeldBy.transform.position, playerHeldBy.previousYRot, index);
-                }
+                ActivateCandy(itemProperties.itemName, pinataCandy);
             }
         }
 
-        private void ActivateCandy(string candyName)
+        public static void ActivateCandy(string candyName, bool pinataCandy)
         {
-            switch (candyName)
+            HUDManager.Instance.UIAudio.PlayOneShot(CandyCrunchsfx, 1f);
+
+            if (!pinataCandy || PlayerAge >= 12)
             {
-                case "Blue Candy":
-                    logger.LogDebug("Candy blue");
-                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyBlueEffects.Value); }
-                    else
+                if (pinataCandy && (int)UnityEngine.Random.Range(0, 101) < configCandyDeathChance.Value)
+                {
+                    localPlayer.KillPlayer(new Vector3(), true, CauseOfDeath.Unknown, 3);
+                    return;
+                }
+
+                if (!pinataCandy || configSecretLab.Value)
+                {
+                    logger.LogDebug("Activating candy effects...");
+
+                    switch (candyName)
                     {
-                        StatusEffectController.Instance.HealPlayer(30, true);
+                        case "Blue Candy":
+                            logger.LogDebug("Candy blue");
+                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyBlueEffects.Value); }
+                            else
+                            {
+                                StatusEffectController.Instance.HealPlayer(30, true);
+                            }
+                            break;
+                        case "Green Candy":
+                            logger.LogDebug("Candy green");
+                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyGreenEffects.Value); }
+                            else
+                            {
+                                StatusEffectController.Instance.StatusNegation(30);
+                                StatusEffectController.Instance.HealthRegen(1, 80);
+                            }
+                            break;
+                        case "Purple Candy":
+                            logger.LogDebug("Candy purple");
+                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyPurpleEffects.Value); }
+                            else
+                            {
+                                StatusEffectController.Instance.DamageReduction(15, 20, true);
+                                StatusEffectController.Instance.HealthRegen(2, 10);
+                            }
+                            break;
+                        case "Red Candy":
+                            logger.LogDebug("Candy red");
+                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyRedEffects.Value); }
+                            else
+                            {
+                                StatusEffectController.Instance.HealthRegen(9, 5);
+                            }
+                            break;
+                        case "Yellow Candy":
+                            logger.LogDebug("Candy yellow");
+                            if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyYellowEffects.Value); }
+                            else
+                            {
+                                StatusEffectController.Instance.RestoreStamina(25);
+                                StatusEffectController.Instance.InfiniteSprint(8);
+                                StatusEffectController.Instance.IncreasedMovementSpeed(8, 2, true, true);
+                            }
+                            break;
+                        case "Pink Candy":
+                            logger.LogDebug("Candy pink");
+                            Landmine.SpawnExplosion(localPlayer.transform.position, true, 3, 3);
+                            break;
+                        case "Rainbow Candy":
+                            logger.LogDebug("Candy rainbow");
+                            StatusEffectController.Instance.HealPlayer(15);
+                            StatusEffectController.Instance.InfiniteSprint(5, true);
+                            StatusEffectController.Instance.bulletProofMultiplier += 1;
+                            StatusEffectController.Instance.StatusNegation(10);
+                            StatusEffectController.Instance.HealPlayer(20, true);
+                            break;
+                        case "Black Candy":
+                            logger.LogDebug("Candy black");
+                            ActivateCandy(CandyNames.Where(x => x != "Black Candy").ToList()[UnityEngine.Random.Range(0, CandyNames.Count - 1)], pinataCandy);
+                            break;
+                        default:
+                            logger.LogDebug("Candy not found");
+                            break;
                     }
-                    break;
-                case "Green Candy":
-                    logger.LogDebug("Candy green");
-                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyGreenEffects.Value); }
-                    else
-                    {
-                        StatusEffectController.Instance.StatusNegation(30);
-                        StatusEffectController.Instance.HealthRegen(1, 80);
-                    }
-                    break;
-                case "Purple Candy":
-                    logger.LogDebug("Candy purple");
-                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyPurpleEffects.Value); }
-                    else
-                    {
-                        StatusEffectController.Instance.DamageReduction(15, 20, true);
-                        StatusEffectController.Instance.HealthRegen(2, 10);
-                    }
-                    break;
-                case "Red Candy":
-                    logger.LogDebug("Candy red");
-                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyRedEffects.Value); }
-                    else
-                    {
-                        StatusEffectController.Instance.HealthRegen(9, 5);
-                    }
-                    break;
-                case "Yellow Candy":
-                    logger.LogDebug("Candy yellow");
-                    if (configEnableCustomStatusEffects.Value) { StatusEffectController.Instance.ApplyCandyEffects(configCandyYellowEffects.Value); }
-                    else
-                    {
-                        StatusEffectController.Instance.RestoreStamina(25);
-                        StatusEffectController.Instance.InfiniteSprint(8);
-                        StatusEffectController.Instance.IncreasedMovementSpeed(8, 2, true, true);
-                    }
-                    break;
-                case "Pink Candy":
-                    logger.LogDebug("Candy pink");
-                    Landmine.SpawnExplosion(playerHeldBy.transform.position, true, 3, 3);
-                    break;
-                case "Rainbow Candy":
-                    logger.LogDebug("Candy rainbow");
-                    StatusEffectController.Instance.HealPlayer(15);
-                    StatusEffectController.Instance.InfiniteSprint(5, true);
-                    StatusEffectController.Instance.bulletProofMultiplier += 1;
-                    StatusEffectController.Instance.StatusNegation(10);
-                    StatusEffectController.Instance.HealPlayer(20, true);
-                    break;
-                case "Black Candy":
-                    logger.LogDebug("Candy black");
-                    ActivateCandy(CandyNames.Where(x => x != "Black Candy").ToList()[UnityEngine.Random.Range(0, CandyNames.Count - 1)]);
-                    break;
-                default:
-                    logger.LogDebug("Candy not found");
-                    break;
+                }
+            }
+            else if (configEnablePinata.Value)
+            {
+                // TODO: Animation for player turning into SCP956 and bones crunching sound effects. Maybe spawn in as scavenger model and play animation to turn into scp956!
+                localPlayer.KillPlayer(new Vector3(), false, CauseOfDeath.Unknown, 3);
+                int index = RoundManager.Instance.currentLevel.Enemies.FindIndex(x => x.enemyType.enemyName == "SCP-956");
+                RoundManager.Instance.SpawnEnemyOnServer(localPlayer.transform.position, localPlayer.previousYRot, index);
             }
         }
     }
 }
-// TODO: Add candy bags to hold candies
 
 // From Secret Lab Wiki:
 // Blue: